+++ database/migrations/2017_12_17_235422_create_reports_table.php
@@ -21,8 +21,8 @@ class CreateReportsTable extends Migration
             $table->text('components');//different for bank and company
             $table->float('stock_price')->nullable();//only for company
             $table->float('dividend')->nullable();
-            $table->integer('profit')->nullable();//only for company
-            $table->integer('loan_total')->nullable();//only for bank
+            $table->bigInteger('profit')->nullable();//only for company
+            $table->bigInteger('loan_total')->nullable();//only for bank
             $table->text('buildings')->nullable();//only for company
             $table->timestamps();
             //This is a migration very specific for this contest. Probably useless later.
